refactor: unify profile iteration and update logic
