+++ src/utils.rs
@@ -2,6 +2,7 @@ mod backend;
 mod config;
 pub mod ipc;
 mod macros;
+mod profile;
 pub mod self_update;
 mod state;
 
@@ -11,3 +12,4 @@ pub(crate) use backend::{BackEnd, ServiceOp};
 pub(crate) mod consts;
 
 pub(crate) use config::BuildConfig;
+pub(crate) use profile::Profile;
