+++ src/utils/config.rs
@@ -5,7 +5,7 @@ use std::path::Path;
 use crate::clash::webapi::local_config::{Basic, Service};
 use serde::{Deserialize, Serialize};
 
-use super::backend::profile::map::ProfileDataBase;
+use crate::utils::profile::map::ProfileDataBase;
 
 #[derive(Debug, Default, Deserialize, Serialize)]
 #[serde(default)]
